fix(migration jobs): Change the experience column type to unsignedBigInteger
The wrong type did not allow the foreign key relationship

+++ database/migrations/2023_11_02_135758_create_jobs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table -> string('company',90) -> nullable(false);
             $table -> string('title',50) -> nullable(false);
             $table -> string('description',120) -> nullable(false);
-            $table -> unsignedInteger('experience') -> nullable(false);
+            $table -> unsignedBigInteger('experience') -> nullable(false);
             $table -> unsignedBigInteger('location') -> nullable(false);
             $table->timestamps();
         });
